+++ public/Application/RemoveFromCartUseCase.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Application;
+
+use Domain\Order;
+use Domain\ShoppingCart;
+
+final readonly class RemoveFromCartUseCase {
+
+}
